Protect JavaDocRootType.getInstance in erlang/Type
Use documentRootType to indirectly get the documentation OrderRootType
in the Erlang SDK Type, so that it works in Small IDEs that have more
than `CLASSES` and `SOURCES` root types installed.
